Added codebook

+++ app/src/main/java/com/ivanb/dictaphone/RecordActivity.kt
@@ -7,13 +7,14 @@ import android.content.Intent
 import android.os.Bundle
 import android.os.CountDownTimer
 import android.text.InputType
-import android.widget.EditText
-import android.widget.ImageView
-import android.widget.TextView
+import android.widget.*
 import androidx.appcompat.app.AppCompatActivity
 import com.ivanb.dictaphone.record.AndroidAudioRecorded
 import org.w3c.dom.Text
+import java.io.BufferedWriter
 import java.io.File
+import java.io.FileWriter
+import java.security.MessageDigest
 
 
 class RecordActivity : AppCompatActivity() {
@@ -77,34 +78,53 @@ class RecordActivity : AppCompatActivity() {
                 val builder: AlertDialog.Builder = AlertDialog.Builder(this)
                 builder.setTitle("Save")
 
-                val input = EditText(this)
+                if(!File(applicationContext.filesDir.toString() + "/codebook.txt").exists()) {
+                    Toast.makeText(this, "You need to add at least 1 code first!", Toast.LENGTH_SHORT).show()
+                } else {
 
-                input.inputType =
-                    InputType.TYPE_CLASS_TEXT
-                builder.setView(input)
+                    val options = getCodes().toTypedArray()
 
-                builder.setPositiveButton("Save"
-                ) { _, _ ->
-                    val pathToSave = applicationContext.filesDir.toString() + '/' + input.text.toString()
-                    val destinationAudioFile = File("$pathToSave.mp3")
-                    audioFile?.copyTo(destinationAudioFile, true)
-                    audioFile?.delete()
+                    val spinner = Spinner(this)
+                    spinner.adapter = ArrayAdapter(this, android.R.layout.simple_spinner_dropdown_item, options)
+                    builder.setView(spinner)
 
-                    val data = Intent()
-                    data.putExtra("filePath", destinationAudioFile.path.toString())
-                    setResult(Activity.RESULT_OK, data)
-                    finish()
+                    builder.setPositiveButton("Save"
+                    ) { _, _ ->
+                        val currentTimeMillis = System.currentTimeMillis()
+                        val timeBytes = currentTimeMillis.toString().toByteArray()
+                        val digest = MessageDigest.getInstance("SHA-256")
+                        val hashBytes = digest.digest(timeBytes)
+                        val hashString = hashBytes.joinToString("") { "%02x".format(it) }
+
+                        val pathToSave = applicationContext.filesDir.toString() + '/' + options[spinner.selectedItemPosition] + '-' + hashString
+                        val destinationAudioFile = File("$pathToSave.mp3")
+                        audioFile?.copyTo(destinationAudioFile, true)
+                        audioFile?.delete()
+
+                        if(descriptionInput.text.toString().isNotEmpty()) {
+                            val destinationDescriptionFile = File("$pathToSave.txt")
+                            val fileWriter = FileWriter(destinationDescriptionFile)
+                            val bufferedWriter = BufferedWriter(fileWriter)
+                            bufferedWriter.write(descriptionInput.text.toString())
+                            bufferedWriter.close()
+                        }
+
+                        val data = Intent()
+                        data.putExtra("filePath", destinationAudioFile.path.toString())
+                        setResult(Activity.RESULT_OK, data)
+                        finish()
+                    }
+
+                    builder.setNegativeButton("Cancel"
+                    ) { dialog, _ ->
+                        dialog.cancel()
+
+                        setResult(Activity.RESULT_CANCELED)
+                        finish()
+                    }
+
+                    builder.show()
                 }
-
-                builder.setNegativeButton("Cancel"
-                ) { dialog, _ ->
-                    dialog.cancel()
-
-                    setResult(Activity.RESULT_CANCELED)
-                    finish()
-                }
-
-                builder.show()
             }
         }
     }
@@ -114,5 +134,13 @@ class RecordActivity : AppCompatActivity() {
         val minutes = seconds / 60
         val remainingSeconds = seconds % 60
         currentTimeTextView.text = "$minutes:${String.format("%02d", remainingSeconds)}"
+    }
+
+    private fun getCodes(): List<String> {
+
+        val codebookFile = File(applicationContext.filesDir.toString() + "/codebook.txt")
+
+        val codes = codebookFile.readLines()
+        return codes.sorted()
     }
 }
